cleaned and untested prototype
