+++ packages/railtracks/src/railtracks/built_nodes/concrete/__init__.py
@@ -2,13 +2,6 @@ __all__ = [
     "StringResponse",
     "StructuredResponse",
     "TerminalLLM",
-    "GuardedTerminalLLM",
-    "GuardedStreamingTerminalLLM",
-    "GuardedStructuredLLM",
-    "GuardedStreamingStructuredLLM",
-    "GuardedToolCallLLM",
-    "GuardedStreamingToolCallLLM",
-    "GuardedStructuredToolCallLLM",
     "StructuredLLM",
     "ToolCallLLM",
     "StructuredToolCallLLM",
@@ -26,15 +19,7 @@ from .chat_tool_call_llm import ChatToolCallLLM
 from .function_base import (
     RTFunction,
 )
-from .guarded_llm import (
-    GuardedStreamingStructuredLLM,
-    GuardedStreamingTerminalLLM,
-    GuardedStreamingToolCallLLM,
-    GuardedStructuredLLM,
-    GuardedStructuredToolCallLLM,
-    GuardedTerminalLLM,
-    GuardedToolCallLLM,
-)
+
 from .response import StringResponse, StructuredResponse
 from .structured_llm_base import StructuredLLM
 from .structured_tool_call_llm_base import StructuredToolCallLLM
