+++ packages/railtracks/src/railtracks/guardrails/llm/__init__.py
@@ -3,7 +3,6 @@ from ._pii.config import PIICustomPattern, PIIEntity, PIIRedactConfig
 from .input.block_text import BlockTextInputGuard
 from .input.length_guard import InputLengthGuard
 from .input.pii_redact import PIIRedactInputGuard
-from .mixin import LLMGuardrailsMixin
 from .output.block_text import BlockTextOutputGuard
 from .output.length_guard import OutputLengthGuard
 from .output.pii_redact import PIIRedactOutputGuard
@@ -14,7 +13,6 @@ __all__ = [
     "BlockTextInputGuard",
     "BlockTextOutputGuard",
     "InputLengthGuard",
-    "LLMGuardrailsMixin",
     "OutputLengthGuard",
     "PIICustomPattern",
     "PIIEntity",
