+++ packages/railtracks/src/railtracks/guardrails/llm/input/length_guard.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from railtracks.guardrails.core.decision import GuardrailDecision
 from railtracks.guardrails.core.event import LLMGuardrailEvent
-from railtracks.guardrails.core.interfaces import InputGuard
+from railtracks.guardrails.llm.concrete import InputGuard
 
 
 class InputLengthGuard(InputGuard):
