+++ packages/railtracks/src/railtracks/guardrails/llm/output/length_guard.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from railtracks.guardrails.core.decision import GuardrailDecision
 from railtracks.guardrails.core.event import LLMGuardrailEvent
-from railtracks.guardrails.core.interfaces import OutputGuard
+from railtracks.guardrails.llm.concrete import OutputGuard
 
 
 class OutputLengthGuard(OutputGuard):
